jobboard/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.2.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
django-extensions==3.2.3
